+++ src/interfaces/modal.ts
@@ -15,6 +15,14 @@ export interface ISymptomModal {
   select: number;
 }
 
-export interface ILoginModal {
+export interface IMainModal {
   closeModal: () => void;
+}
+
+export interface ILoginModal {
+  handleLogin: () => void;
+  closeModal: () => void;
+  handleContinue: () => void;
+  title: ReactNode;
+  continueText: string;
 }
